+++ conda/recipes/libcustrings/build.sh
@@ -1,17 +1,10 @@
 #!/usr/bin/env bash
 
-CMAKE_COMMON_VARIABLES=" -DCMAKE_INSTALL_PREFIX=$PREFIX -DCMAKE_BUILD_TYPE=Release"
+# Copyright (c) 2018-2019, NVIDIA CORPORATION.
 
 # show environment
 printenv
 # Cleanup local git
 git clean -xdf
-# Change directory for build process
-cd cpp
-# Use CMake-based build procedure
-mkdir build
-cd build
-# configure
-cmake $CMAKE_COMMON_VARIABLES ..
-# build cpp
-make -j${PARALLEL_LEVEL} VERBOSE=1 install
+# This assumes the script is executed from the root of the repo directory
+./build.sh -v libcustrings